# schemas/openapi-commercial-v1.in.yaml
openapi: "3.0.3"
info:
  title: AgentSkeptic commercial license API
  version: "1.0.0"
  contact:
    url: __DISTRIBUTION_CONTACT_URL__
  x-agentskeptic-distribution:
    repository: __DISTRIBUTION_REPO_URL__
    npmPackage: __DISTRIBUTION_NPM_URL__
    openApi: __OPENAPI_SELF_URL__
  description: "__IDENTITY_ONE_LINER__\n\nMachine-readable contract for license preflight used by the published npm CLI.\nBase URL is your deployed app origin (same as NEXT_PUBLIC_APP_URL).\n\nEvery path in this document returns an `x-request-id` response header on all status codes (echo a valid client `x-request-id` when supplied; otherwise server-generated). Non-2xx JSON bodies follow RFC 7807-style Problem Details (`type`, `title`, `status`, `detail`, optional `code`, `instance`) unless noted; `POST /api/v1/usage/reserve` denials additionally include legacy `allowed`, `code`, `message`, and optional `upgrade_url` for backward compatibility.\n"
...
servers:
  - url: __SERVERS_ORIGIN__
    description: Replace with your deployment origin
paths:
  /api/v1/usage/reserve:
    post:
      operationId: reserveUsage
      summary: Reserve one verification or enforcement run (license preflight)
      description: >
        Called by the commercial CLI before contract verify, quick verify, or enforce.
        Requires Bearer API key. Body run_id must be idempotent per key.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReserveRequest"
      responses:
        "200":
          description: Allowed; quota consumed or duplicate run_id returned without double-charge
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveAllowed"
        "400":
          description: Bad request (missing/invalid body, skewed issued_at)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
        "401":
          description: Invalid or revoked API key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
        "403":
          description: Quota exceeded or entitlement denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
        "503":
          description: Server error during reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReserveError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
  schemas:
    ProblemDetails:
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        code:
          type: string
...
    ReserveRequest:
      type: object
      required:
        - run_id
        - issued_at
      properties:
        run_id:
          type: string
          maxLength: 256
        issued_at:
          type: string
          format: date-time
          description: ISO8601; server rejects if skew from now exceeds 300 seconds
        intent:
          type: string
          enum: [verify, enforce]
          default: verify
          description: >
            enforce requires a paid plan (not Starter) with active subscription; verify is allowed
            for Starter (free included quota) and for paid plans with active subscription; overage
            is permitted past included for paid plans with allowOverage in plan catalog
    ReserveAllowed:
      type: object
      required:
        - allowed
        - plan
        - limit
        - used
        - included_monthly
        - overage_count
      properties:
        allowed:
          type: boolean
          enum: [true]
        plan:
          type: string
          enum: [starter, individual, team, business, enterprise]
        limit:
          type: integer
          description: Included verifications in the current billing month (not enterprise unlimited sentinel)
        used:
          type: integer
          description: Total verifications counted this month for this API key
        included_monthly:
          oneOf:
            - type: integer
            - type: "null"
          description: Same as limit for finite plans; null for enterprise unlimited
        overage_count:
          type: integer
          description: max(0, used - included_monthly) when the plan has a finite cap
    ReserveError:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          required: [allowed, code, message]
          properties:
            allowed:
              type: boolean
              enum: [false]
            code:
              type: string
              description: >
                Examples BAD_REQUEST, INVALID_KEY, QUOTA_EXCEEDED, ENFORCEMENT_REQUIRES_PAID_PLAN,
                VERIFICATION_REQUIRES_SUBSCRIPTION, SUBSCRIPTION_INACTIVE, BILLING_PRICE_UNMAPPED, SERVER_ERROR
            message:
              type: string
            upgrade_url:
              type: string
              format: uri
              description: Present for some 403 entitlement responses
...